refactor: extract top actions into analyzer/actions.rs (#190)
Moves TopAction + compute_top_actions (and the short_dir / short_file
formatting helpers used only there) out of analyzer/mod.rs.
